Modify usage of template helper
